SVN r4541 (0.6.x)

+++ common/i_net.cpp
@@ -4,7 +4,7 @@
 // $Id$
 //
 // Copyright (C) 1993-1996 by id Software, Inc.
-// Copyright (C) 2006-2013 by The Odamex Team.
+// Copyright (C) 2006-2014 by The Odamex Team.
 //
 // This program is free software; you can redistribute it and/or
 // modify it under the terms of the GNU General Public License
@@ -1069,7 +1069,7 @@ void InitNetCommon(void)
 //
 bool NetWaitOrTimeout(size_t ms)
 {
-	struct timeval timeout = {0, (1000*ms) + 1};
+	struct timeval timeout = {0, int(1000*ms) + 1};
 	fd_set fds;
 
 	FD_ZERO(&fds);
